fix(summary): validate form when preview is enable
Before: when preview is enable and we insert an error in the form the preview is still enable.
Solution: validate the form on preview button click and disable preview if there is an error.
After: when preview is enable and we insert an error in the form the preview is disable on preview button click

Close #202